refactor: cookies
