+++ internal/auth/controllers/session_login.go
@@ -4,7 +4,7 @@ import (
 	"net/http"
 
 	"github.com/indeedhat/barista/internal/auth"
-	"github.com/indeedhat/barista/internal/auth/cookie"
+	"github.com/indeedhat/barista/internal/cookie"
 	"github.com/indeedhat/barista/internal/server"
 	"github.com/indeedhat/barista/internal/ui"
 )
@@ -66,14 +66,7 @@ func (c Controller) Login(rw http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	http.SetCookie(rw, &http.Cookie{
-		Name:     cookie.SessionKey,
-		Value:    jwt,
-		HttpOnly: true,
-		Domain:   r.URL.Host,
-		Path:     "/",
-		MaxAge:   86400 * 30,
-	})
+	cookie.Set(rw, r, cookie.SessionKey, jwt)
 
 	ui.Redirect(rw, "/")
 }
