--- internal/auth/cookie/cookie.go
@@ -1,3 +1,0 @@
-package cookie
-
-const SessionKey = "bs"
